+++ src/components/input/input-interface.ts
@@ -1,6 +1,6 @@
-// From: https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/autocomplete
 import { WcsSize } from '../../shared-types';
 
+// From: https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/autocomplete
 export type AutocompleteTypes = (
     | 'on' | 'off' | 'name' | 'honorific-prefix' | 'given-name' | 'additional-name' | 'family-name' | 'honorific-suffix'
     | 'nickname' | 'email' | 'username' | 'new-password' | 'current-password' | 'one-time-code' | 'organization-title' | 'organization'
@@ -12,14 +12,8 @@ export type AutocompleteTypes = (
 
 export type TextFieldTypes = 'date' | 'email' | 'number' | 'password' | 'search' | 'tel' | 'text' | 'url' | 'time' | 'week' | 'month' | 'datetime-local' | 'file';
 
-export const WcsInputSizeValues = ['s', 'm', 'l'] as const; // as const keyword is used to infer and preserve the exact literal values of an array or object.
 
-export type WcsInputSize = Extract<WcsSize, typeof WcsInputSizeValues[number]>;
-
-export function isWcsInputSize(size: string): size is WcsInputSize {
-    // @ts-ignore : ignore size type, as it is checked with WcsInputSizeValues
-    return WcsInputSizeValues.includes(size);
-}
+export type WcsInputSize = Extract<WcsSize, 's' | 'm' | 'l'>;
 
 export type WcsInputAutocorrect = 'on' | 'off';