+++ src/components/modal/modal-interface.ts
@@ -1,1 +1,3 @@
-export type ModalSize = 's' | 'm' | 'l' | 'xl';
+import { WcsSize } from "../../shared-types";
+
+export type ModalSize = Extract<WcsSize, 's' | 'm' | 'l' | 'xl'>;